Relocate namspaces and preferences to utils
