+++ src/main/java/net/cytonic/cytosis/commands/IgnoreChatChannelCommand.java
@@ -2,8 +2,8 @@ package net.cytonic.cytosis.commands;
 
 import com.google.gson.JsonObject;
 import net.cytonic.cytosis.Cytosis;
-import net.cytonic.cytosis.data.enums.CytosisNamespaces;
-import net.cytonic.cytosis.data.enums.CytosisPreferences;
+import net.cytonic.cytosis.utils.CytosisNamespaces;
+import net.cytonic.cytosis.utils.CytosisPreferences;
 import net.cytonic.enums.ChatChannel;
 import net.minestom.server.command.builder.Command;
 import net.minestom.server.command.builder.arguments.ArgumentType;
