+++ src/main/java/net/cytonic/cytosis/events/ServerEventListeners.java
@@ -3,10 +3,10 @@ package net.cytonic.cytosis.events;
 import net.cytonic.cytosis.Cytosis;
 import net.cytonic.cytosis.commands.server.TPSCommand;
 import net.cytonic.cytosis.config.CytosisSettings;
-import net.cytonic.cytosis.data.enums.CytosisPreferences;
 import net.cytonic.cytosis.data.enums.NPCInteractType;
 import net.cytonic.cytosis.logging.Logger;
 import net.cytonic.cytosis.npcs.NPC;
+import net.cytonic.cytosis.utils.CytosisPreferences;
 import net.cytonic.enums.ChatChannel;
 import net.minestom.server.entity.GameMode;
 import net.minestom.server.entity.Player;
